requirements.txt
langchain==0.3.25  
langgraph==0.4.8  
langsmith==0.2.22
openai==1.86.0  
google-generativeai==0.8.5  
python-dotenv==1.1.0  
faiss-cpu==1.11.0  
pytest==8.4.0  
langchain_openai==0.3.22